requirements.txt
# requirements.txt
docker
requests
google-generativeai
python-dotenv
openai
pytest
beautifulsoup4
PyGithub